small fixes, memory footprint improvement

+++ trace-cobertura/src/main/java/se/de/hu_berlin/informatik/spectra/provider/tracecobertura/infrastructure/comptrace/ReplaceableCloneableIterator.java
@@ -7,6 +7,6 @@ public interface ReplaceableCloneableIterator<T> extends CloneableIterator<T> {
 
 	public T processNextAndReplaceWithResult(Function<T,T> function);
 
-	public void setToPosition(int i);
+	public void setToPosition(long i);
 	
 }
