+++ trace-cobertura/src/main/java/se/de/hu_berlin/informatik/spectra/provider/tracecobertura/infrastructure/comptrace/TraceIterator.java
@@ -187,7 +187,7 @@ public class TraceIterator<T> implements ReplaceableCloneableIterator<T> {
 	}
 
 	@Override
-	public void setToPosition(int i) {
+	public void setToPosition(long i) {
 		throw new UnsupportedOperationException();
 	}
